requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
openai>=1.58.1
tavily-python>=0.5.0
httpx>=0.28.0
asyncpg>=0.30.0
pydantic-settings>=2.7.0
sse-starlette>=2.2.0
python-dotenv>=1.0.0
trafilatura>=2.0.0
markitdown>=0.1.1
beautifulsoup4>=4.12.0
chromadb>=0.5.23
sentence-transformers>=3.3.1